Change to double for total mod size for modpacks larger than 2gb

+++ FFXIV_TexTools2/Views/ImportModPack.xaml.cs
@@ -80,7 +80,7 @@ namespace FFXIV_TexTools2.Views
                 Debug.WriteLine(ex.StackTrace);
             }
 
-            float totalModSize = 0;
+            double totalModSize = 0;
 
             foreach(var mpi in mpjList)
             {
